1. The last digit

+++ 0x01-variables_if_else_while/1-last_digit.c
@@ -6,14 +6,14 @@
 /* betty style doc for function main goes there */
 int main(void)
 {
-	long n, last;
+	int n, last;
 
 	srand(time(0));
 	n = rand() - RAND_MAX / 2 ;
 	last = n % 10;
 	if (last > 5)
 	{
-		printf("Last digit of %d is %d and is greater than 5", n, last);
+		printf("Last digit of %d is %d and is greater than 5",n, last);
 	}
 	else
 	{
